Fix 4459 Change of the default painter behavior.
