+++ org.mwc.debrief.legacy/src/Debrief/GUI/Tote/Painters/SnailPainter.java
@@ -246,7 +246,6 @@ import Debrief.Wrappers.Track.LightweightTrackWrapper;
 import Debrief.Wrappers.Track.TrackSegment;
 import Debrief.Wrappers.Track.TrackWrapper_Support.SegmentList;
 import MWC.Algorithms.PlainProjection;
-import MWC.GUI.BaseLayer;
 import MWC.GUI.CanvasType;
 import MWC.GUI.Editable;
 import MWC.GUI.Layer;
@@ -353,139 +352,6 @@ public class SnailPainter extends TotePainter
 	public final void setVectorStretch(final double val)
 	{
 		_mySnailPlotter.setVectorStretch(val);
-	}
-
-	/**
-	 * ok, get the watchables for this layer
-	 * 
-	 * @param theData
-	 *          the layer to extract the watchables for
-	 * @return the set of watchable items in this layer
-	 */
-	public static Vector<Plottable> getWatchables(final Layer thisLayer)
-	{
-		// get the output ready
-		final Vector<Plottable> res = new Vector<Plottable>(0, 1);
-
-		// is this layer visible?
-		if (thisLayer.getVisible())
-		{
-			// is this a watchable?
-			if (thisLayer instanceof WatchableList)
-			{
-				// just double-check this isn't a buoy-pattern, we don't want to display
-				// them
-				if (thisLayer instanceof Editable.DoNotHighlightMe)
-				{
-					// ignore it, we don't want to plot it.
-				}
-				else
-				{
-
-					res.addElement(thisLayer);
-
-					// just have a look if it's a track - if so we want to add it's
-					// sensors
-					if (thisLayer instanceof TrackWrapper)
-					{
-						final TrackWrapper trw = (TrackWrapper) thisLayer;
-
-						// first plot the sensors
-						final BaseLayer sensorsLayer = trw.getSensors();
-						if (sensorsLayer.getVisible())
-						{
-							final Enumeration<Editable> sensors = sensorsLayer.elements();
-							if (sensors != null)
-							{
-								while (sensors.hasMoreElements())
-								{
-									final SensorWrapper sw = (SensorWrapper) sensors
-											.nextElement();
-									// just check if it's visible
-									if (sw.getVisible())
-									{
-										res.add(sw);
-									}
-								}
-							}
-						}
-
-						// now the TMA solutons
-						final BaseLayer tuaLayer = trw.getSolutions();
-						if (tuaLayer.getVisible())
-						{
-							final Enumeration<Editable> solutions = tuaLayer.elements();
-							if (solutions != null)
-							{
-								while (solutions.hasMoreElements())
-								{
-									final TMAWrapper sw = (TMAWrapper) solutions.nextElement();
-									// just check if it's visible
-									if (sw.getVisible())
-									{
-										res.add(sw);
-									}
-								}
-							}
-						}
-					}
-				}
-			}
-			else
-			{
-				final Enumeration<Editable> iter = thisLayer.elements();
-				while (iter.hasMoreElements())
-				{
-
-					final Editable thisE = iter.nextElement();
-					if (thisE instanceof Plottable)
-					{
-						final Plottable p = (Plottable) thisE;
-						if (p instanceof WatchableList)
-						{
-							// look at the date date
-							final WatchableList wl = (WatchableList) p;
-							final HiResDate startDTG = wl.getStartDTG();
-
-							// is it a real date?
-							if (startDTG != null)
-							{
-								// yup, add to list
-								res.addElement(p);
-							}
-						}
-					}
-				}
-			}
-		}
-
-		return res;
-	}
-
-	/**
-	 * ok, get the watchables for this set of layers
-	 * 
-	 * @param theData
-	 *          the layers to extract the watchables for
-	 * @return the set of watchable items in these layers
-	 */
-	public static Vector<Plottable> getWatchables(final Layers theData)
-	{
-		final Vector<Plottable> res = new Vector<Plottable>(0, 1);
-		// step through the layers
-		final int cnt = theData.size();
-		for (int i = 0; i < cnt; i++)
-		{
-			// ok, do this layer
-			final Layer thisLayer = theData.elementAt(i);
-
-			// get the watchables from this layer
-			final Vector<Plottable> newElements = getWatchables(thisLayer);
-
-			// and add to our growing total
-			res.addAll(newElements);
-		}
-		return res;
 	}
 
 	/**
